+++ src/game/Objects/Player.cpp
@@ -2653,7 +2653,7 @@ bool Player::IsAllowedWhisperFrom(ObjectGuid guid) const
     return false;
 }
 
-bool Player::IsGroupVisibleFor(Player* p) const
+bool Player::IsGroupVisibleFor(Player const* p) const
 {
     switch (sWorld.getConfig(CONFIG_UINT32_GROUP_VISIBILITY))
     {
@@ -18068,7 +18068,7 @@ bool Player::CanJoinToBattleground() const
     return true;
 }
 
-bool Player::IsVisibleInGridForPlayer(Player* pl) const
+bool Player::IsVisibleInGridForPlayer(Player const* pl) const
 {
     // gamemaster in GM mode see all, including ghosts
     if (pl->IsGameMaster() && GetSession()->GetSecurity() <= pl->GetSession()->GetSecurity())
